rename: app.js 위치 루트 -> Compnent/ 변경
